Extract logic to separate files
